requirements.txt
alembic==1.7.6
asn1crypto==0.24.0
Babel==2.6.0
blinker==1.4
cffi==1.15.0
click==8.0.4
cryptography==36.0.2
dominate==2.3.1
Flask==2.0.3
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
idna==2.7
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
pycparser==2.21
PyJWT==2.3.0
PyMySQL==0.9.2
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
six==1.11.0
SQLAlchemy==1.4.31
visitor==0.1.3
Werkzeug==2.0.3
WTForms==2.2.1



